requirements.txt
# clip
numpy
torch
transformers
requests
torchvision
flask
werkzeug
pillow
requests
clip @ git+https://github.com/openai/CLIP.git
django>=4.0